% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpact.R
\name{aggregate_pop}
\alias{aggregate_pop}
\title{Given population data for each age group and coverage for a range of ages e.g. 1 to 100
return coverage data with total population over this range}
\usage{
aggregate_pop(coverage, population)
}
\arguments{
\item{coverage}{Tibble of coverage data}

\item{population}{Tibble of population data}
}
\value{
Tibble of coverage data with aggregated population
}
\description{
Given population data for each age group and coverage for a range of ages e.g. 1 to 100
return coverage data with total population over this range
}
\keyword{internal}
