% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpact.R
\name{calculate_impact}
\alias{calculate_impact}
\title{Calculate impact - requires a DB connection to montagu.}
\usage{
calculate_impact(
  con,
  method,
  touchstone,
  modelling_group,
  disease,
  focal_scenario_type,
  focal_vaccine_delivery = NULL,
  baseline_scenario_type,
  baseline_vaccine_delivery = NULL,
  burden_outcomes = c("deaths", "cases", "dalys"),
  countries = NULL,
  is_under5 = FALSE,
  vaccination_years = 2000:2030
)
}
\arguments{
\item{con}{Connection to database.}

\item{method}{Impact method to use one of calendar_year, birth_year,
yov_activity_type, yov_birth_cohort.}

\item{touchstone}{The montagu touchstone to calculate impact for. Either
touchstone ID or touchstone name.}

\item{modelling_group}{The modelling group to calculate impact for.}

\item{disease}{The disease to calculate impact for.}

\item{focal_scenario_type}{The focal scenario scenario type e.g. "default"}

\item{focal_vaccine_delivery}{The focal vaccine delivery methods. This should
be a list of lists or NULL if scenario type is novac. Each element of first
list needs to specify the vaccination and activity type. e.g.
list(
  list(
    vaccine = "HepB",
    activity_type = "routine"
  ),
  list(
    vaccine = "HepB_BD",
    activity_type = "routine"
  )
 )}

\item{baseline_scenario_type}{The baseline scenario scenario type e.g.
"novac"}

\item{baseline_vaccine_delivery}{Like `focal_vaccine_delivery` this should be
a list of lists, each element containing vaccination and activity type. If
`baseline_vaccine_delivery` is `novac` then this should be NULL.}

\item{burden_outcomes}{List of burden outcomes, defaults to "deaths", "cases"
and "dalys".}

\item{countries}{Vector of countries to get impact for. If NULL then impact
calculated for all countries.}

\item{is_under5}{If TRUE then only include data for age under 5, otherwise
calculate impact for all ages}

\item{vaccination_years}{Years of vaccination of interest, only used for
year of vaccination (yov) methods}
}
\value{
Impact for this set of parameters.
}
\description{
This depends on the DB format of VIMC and so is for internal use only.
}
