% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpact.R
\name{calculate_impact_from_recipe}
\alias{calculate_impact_from_recipe}
\title{Calculate impact from a recipe - requires a DB connection to montagu.}
\usage{
calculate_impact_from_recipe(
  con,
  recipe_path,
  method,
  countries = NULL,
  is_under5 = FALSE,
  vaccination_years = 2000:2030
)
}
\arguments{
\item{con}{Connection to database}

\item{recipe_path}{Path to file containing recipe for burden outcome
calculation. For more details see \code{vignette} vignette.}

\item{method}{Impact method to use one of calendar_year, birth_year,
yov_activity_type, yov_birth_cohort.}

\item{countries}{Vector of countries to get impact for. If NULL then impact
calculated for all countries.}

\item{is_under5}{If TRUE then only include data for age under 5, otherwise
calculate impact for all ages}

\item{vaccination_years}{Years of vaccination of interest, only used for
year of vaccination (yov) methods}
}
\value{
The impact for each row in the recipe
}
\description{
This depends on the DB format of VIMC and so is for internal use only.
}
