% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonly_used_db_data.R
\name{cohort_deaths_all_cause}
\alias{cohort_deaths_all_cause}
\title{Extract demographic data - all couse mortality, live birth and under5 mortality rate}
\usage{
cohort_deaths_all_cause(con, touchstone_pop, cohorts, under_5 = TRUE)
}
\arguments{
\item{con}{Database connection}

\item{touchstone_pop}{Demography touchstone}

\item{cohorts}{birth cohorts for which to extract data}

\item{under_5}{whether constrain to under5 mortality}
}
\description{
Extract demographic data - all couse mortality, live birth and under5 mortality rate
}
