% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonly_used_db_data.R
\name{extract_vaccination_history}
\alias{extract_vaccination_history}
\title{Generate target, fvps, coverage for a touchstone
This function pulls vaccination and population data from Montagu database, and transform them into target, coverage and fvps.
There are two modes of pulling population data: 1.) through touchstone_pop parameter 2.) through demographic_source parameter
touchstone_pop is one of touchstone_demographic_dataset.touchstone; this parameter is null-able, when it is null, touchstone_cov is used
demographic_source is one of demographic_source.code; by default it is null, when it is not null, it rulls out touchstone_pop parameter}
\usage{
extract_vaccination_history(
  con,
  touchstone_cov = "201710gavi",
  touchstone_pop = NULL,
  year_min = 2000,
  year_max = 2100,
  vaccine_to_ignore = c("DTP3", "HepB_BD_home", "none"),
  disease_to_extract = NULL,
  countries_to_extract = NULL,
  gavi_support_levels = c("with", "bestminus"),
  scenario_type = "default",
  external_population_estimates = NULL,
  full_description = FALSE,
  demographic_source = NULL,
  coverage_scenario_type = NULL,
  db_memory = NULL
)
}
\arguments{
\item{con}{Datebase connection}

\item{touchstone_cov}{Coverage touchstone}

\item{touchstone_pop}{touchstone_demographic_dataset.touchstone}

\item{year_min}{extract data from year_min}

\item{year_max}{extract data to year_max}

\item{vaccine_to_ignore}{Ignore defined vaccines}

\item{disease_to_extract}{extract data for specific diseases}

\item{countries_to_extract}{extract data for specific countries}

\item{gavi_support_levels}{gavi support levels}

\item{scenario_type}{scenario type}

\item{external_population_estimates}{The rationales are 1. we can use external population estimates if any and if necessary;
2. demographic uncertainty not only affects models, but also FVPs. If we are to conduct sensitivity analysis on impact_by_year_of_vaccination, we need to vary population input for adjusting FVPs.}

\item{full_description}{TRUE if including scenario_descriptions (coverage estimates will be duplicated for scenarios); and FALSE if only providing coverage estimates}

\item{demographic_source}{Demographic_source.code}

\item{coverage_scenario_type}{Coverage scenario type. This is particularly useful for a coverage touchstone. Montagu coverage.name follows <disease>:<vaccine>,<activity_type>,<gavi_support_level>:<coverage_scenario_type>
It is NULL by default. When it is not null, only pulls coverage.name that contains specified pattern.}

\item{db_memory}{if not null, coverage is pulled from in-memory db; this is used from 202507july onwards, as we are moving away from db imports}
}
\description{
Replace jenner:::fix_covreage_fvps() function
}
