% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpact.R
\name{get_burden_estimate_set_ids}
\alias{get_burden_estimate_set_ids}
\title{Get burden estimate set ids}
\usage{
get_burden_estimate_set_ids(
  con,
  baseline_scenario_type,
  baseline_scenario,
  focal_scenario_type,
  focal_scenario,
  touchstone,
  modelling_group,
  disease
)
}
\arguments{
\item{con}{DB connection}

\item{baseline_scenario_type}{Baseling scenario type, e.g. default, novac}

\item{baseline_scenario}{Scenario specification a string with form e.g.
scenario_type-vaccine1-activity_type1;scenario_type-vaccine2-activity_type2}

\item{focal_scenario_type}{Focal scenario type e.g. default, novac}

\item{focal_scenario}{Scenario specification a string with form e.g.
scenario_type-vaccine1-activity_type1;scenario_type-vaccine2-activity_type2}

\item{touchstone}{The touchstone to get data for}

\item{modelling_group}{The modelling group to get data for}

\item{disease}{The disease to get data for}
}
\value{
Tibble containing the burden estimate set ids which match the
  baseline and focal scenarios
}
\description{
We need to locate unique ID for burden_estimate_set for
this combination of touchstone, modelling_group, disease, vaccine, etc.
There might be multiple scenarios which match the
touchstone, modelling_group, disease, vaccine and activity_type
e.g. default HepB-routine vs default HepB_BD-routine & HepB-routine
We need to go burden-estimate-set-wise and identify which scenario matches
the vaccine_delivery we are interested in.
}
\keyword{internal}
