% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpact.R
\name{get_burden_outcome_ids}
\alias{get_burden_outcome_ids}
\title{Get burden outcome code and ID from a set of codes}
\usage{
get_burden_outcome_ids(con, burden_outcomes)
}
\arguments{
\item{con}{DB connection}

\item{burden_outcomes}{Burden outcome codes to lookup in the database}
}
\value{
dbplyr lazy db connection containing burden_outcome ids and codes
}
\description{
e.g. from "cases" return data frame of id - 1 and code - cases
}
\keyword{internal}
