% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_country}
\alias{get_country}
\title{Get country table}
\usage{
get_country(con)
}
\arguments{
\item{con}{DB connection}
}
\value{
dbplyr lazy db connection to country table
}
\description{
Get country table
}
\keyword{internal}
