% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_coverage}
\alias{get_coverage}
\title{Get coverage table}
\usage{
get_coverage(con)
}
\arguments{
\item{con}{DB connection}
}
\value{
dbplyr lazy db connection to coverage table
}
\description{
Get coverage table
}
\keyword{internal}
