% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_coverage_set_by_touchstone}
\alias{get_coverage_set_by_touchstone}
\title{Get coverage set table}
\usage{
get_coverage_set_by_touchstone(con, touchstone)
}
\arguments{
\item{con}{DB connection}
}
\value{
dbplyr lazy db connection to coverage_set table
}
\description{
Get coverage set table
}
\keyword{internal}
