% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_gender}
\alias{get_gender}
\title{Get gender table}
\usage{
get_gender(con)
}
\arguments{
\item{con}{DB connection}
}
\value{
dbplyr lazy db connection to gender table
}
\description{
Get gender table
}
\keyword{internal}
