% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpact.R
\name{get_impact_for_burden_estimate_set}
\alias{get_impact_for_burden_estimate_set}
\title{Get burden estimate data from db for a burden estimate set and outcome}
\usage{
get_impact_for_burden_estimate_set(
  con,
  burden_estimate_sets,
  outcomes,
  countries,
  is_under5
)
}
\arguments{
\item{con}{DB connection}

\item{burden_estimate_sets}{Burden estimate set ids to get data from}

\item{outcomes}{Burden outcome ids to get data for}

\item{countries}{Optional filter of countries to pull data for}

\item{is_under5}{If TRUE only gets burden estimate data for under 5
ages}
}
\value{
dbplyr lazy db connection containing the raw impact data
}
\description{
Given one or more burden estimate set ids and outcome ids return the
raw impact data fr that burden estimate. Optionally filter to get only
under 5 ages or for a particular country
}
\keyword{internal}
