% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonly_used_db_data.R
\name{get_population}
\alias{get_population}
\title{Extract demographic data}
\usage{
get_population(
  con,
  touchstone_pop = "201710gavi-5",
  demographic_statistic = "int_pop",
  gender = "Both",
  country_ = NULL,
  year_ = NULL,
  age_ = NULL
)
}
\arguments{
\item{con}{Datebase connection}

\item{touchstone_pop}{Demography touchstone}

\item{gender}{Gender codes - "Male", "Female", "Both"}

\item{country_}{All countries if NULL. Or specify a vector of countries}

\item{year_}{All years if NULL. Or specify a vector of years}

\item{age_}{All age groups if NULL. Or specify a vector of age groups}
}
\description{
Extract demographic data
}
