% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpact.R
\name{get_population_dplyr}
\alias{get_population_dplyr}
\title{Get population data for a particular touchstone}
\usage{
get_population_dplyr(con, touchstone, countries = NULL, years = NULL)
}
\arguments{
\item{con}{DB connection}

\item{countries}{Optional vector of countries to filter data}

\item{years}{Option vector of years to filter data}

\item{touchtonse}{Touchstone to get data for}
}
\value{
Tibble of population data
}
\description{
Pull "int_pop" demographic statistic from db for a particular touchstone.
Optionally filter on country and years.
}
\keyword{internal}
