% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_scenario}
\alias{get_scenario}
\title{Get scenario table}
\usage{
get_scenario(con)
}
\arguments{
\item{con}{DB connection}
}
\value{
dbplyr lazy db connection to scenario table
}
\description{
Get scenario table
}
\keyword{internal}
