% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\name{get_scenario_description}
\alias{get_scenario_description}
\title{Get scenario description table}
\usage{
get_scenario_description(con)
}
\arguments{
\item{con}{DB connection}
}
\value{
dbplyr lazy db connection to scenario description table
}
\description{
Get scenario description table
}
\keyword{internal}
