% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stochastic.R
\name{get_stochastic_data}
\alias{get_stochastic_data}
\title{Get data from stochastic tables with specified grouping and filters}
\usage{
get_stochastic_data(
  annex,
  table,
  groups = c("disease", "country", "year"),
  filters = NULL
)
}
\arguments{
\item{annex}{Connection to annex db}

\item{table}{One of cross_all, cross_under5, cohort_all or cohort_under5}

\item{groups}{Categories to group by for aggregating in query}

\item{filters}{Filters to apply before aggregation}
}
\value{
Mean, 0.025 and 0.975 quantiles for deaths_default, deaths_novac.
deaths_impact, dalys_default, dalys_novac, dalys_impact with specified
groupings.
}
\description{
Can query from cross_all, cross_under5, cohort_all and cohort_under5
}
