% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonly_used_db_data.R
\name{get_touchstone}
\alias{get_touchstone}
\title{Determine touchstone}
\usage{
get_touchstone(con, touchstone_name)
}
\arguments{
\item{con}{Database connection.}

\item{touchstone_name}{touchstone_name}
}
\description{
Find latest touchstone given touchstone_name
}
