% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonly_used_db_data.R
\name{get_touchstone_id}
\alias{get_touchstone_id}
\title{Determine touchstone id}
\usage{
get_touchstone_id(con, touchstone)
}
\arguments{
\item{con}{Database connection.}

\item{touchstone}{touchstone name or id}
}
\description{
Find latest touchstone if touchstone name is provided
}
