% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_central.R
\name{impact_by_birth_year}
\alias{impact_by_birth_year}
\title{Calculate impact by birth year (lifetime impact)}
\usage{
impact_by_birth_year(baseline_impact, focal_impact)
}
\arguments{
\item{baseline_impact}{Data frame of baseline impact data this needs to have
columns country, burden_outcome, year, age, value}

\item{focal_impact}{Data frame of focal impact data this needs to have
columns country, burden_outcome, year, age, value}
}
\value{
Vaccine impact by country and birth year for burden outcomes as a
data frame with columns country, year, burden_outcome and impact
}
\description{
The birth year method accounts for the long-term impact accrued over the
lifetime of a particular birth cohort. The duration of modelling needs to
be appropriate to the pathogen of interest as in some cases, such as HepB,
disease occurs later in life. For example if we model vaccination for birth
cohorts born from 2000 to 2030 and model disease burden until 2100 we do
not account for the vaccine impact for those born in 2030 once they are
over 70 years old. The method also does not specifically account for the
impact of vaccinating a cohort outside the cohort vaccinated (e.g.
because of herd protection).
}
