% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_central.R
\name{impact_by_year_of_vaccination_birth_cohort}
\alias{impact_by_year_of_vaccination_birth_cohort}
\title{Calculate impact by year of vaccination: birth cohort}
\usage{
impact_by_year_of_vaccination_birth_cohort(
  baseline_burden,
  focal_burden,
  fvps,
  vaccination_years
)
}
\arguments{
\item{baseline_burden}{Data frame of baseline burden data this needs to have
columns country, burden_outcome, vaccine_delivery, year, age, value}

\item{focal_burden}{Data frame of focal burden data this needs to have
columns country, burden_outcome, vaccine_delivery, year, age, value}

\item{fvps}{Data frame of FVPs (fully vaccinated persons) needs to have
columns country, year, activity_type, fvps other columns can be included
and will be aggregated over.}

\item{vaccination_years}{Years of vaccination of interest.}
}
\value{
Vaccine impact by country, activity type (routine or campaign), year
and burden outcome
}
\description{
Impact by year of vaccination with impact ratio stratified by birth cohort.
Stratifying impact ratio by birth cohort aims to catch temporal changes in
transmission or healthcare.
}
\details{
This method is invariant to activity type. Vaccine effect is assumed to vary
over time through birth cohorts. This means that rather than averaging the
effect of vaccination over time, we account for the variation in transmission
and health of the population. This influences how one year's vaccination may
work compared to another. For example, if therapeutic treatments for a
disease improve over time, we may expect the impact of vaccination in 2050
to be less than that now as the population is generally healthier.
}
