% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_central.R
\name{impact_by_year_of_vaccination_cohort_perspective}
\alias{impact_by_year_of_vaccination_cohort_perspective}
\title{Calculate impact by year of vaccination cohort perspective}
\usage{
impact_by_year_of_vaccination_cohort_perspective(
  raw_impact,
  fvps,
  vaccination_years
)
}
\arguments{
\item{raw_impact}{Data frame of raw impact data this needs to have
columns country, value, burden_outcome and either year & age or birth_cohort}

\item{fvps}{Data frame of fully vaccination person data with columns
country, fvps and either year & age or birth_cohort}

\item{vaccination_years}{Years of vaccination of interest}
}
\value{
Impact ratio by country, birth cohort and burden outcome
}
\description{
This will calculate the impact by year of vaccination by country, birth
cohort and burden outcome for a single disease and vaccine.
}
\details{
This can take data either by vaccination year and age at vaccination or
by birth cohort year.
}
