% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpact.R
\name{impact_ratio_parameters}
\alias{impact_ratio_parameters}
\title{Create parameters for impact calculations}
\usage{
impact_ratio_parameters(
  method,
  fvps,
  countries,
  disease,
  outcome,
  by_country = TRUE,
  cohorts = 2000:2030,
  present = 2019
)
}
\arguments{
\item{method}{Method to use - either cohort or standard}

\item{fvps}{Data frame of FVP information}

\item{countries}{Data frame of country information}

\item{disease}{Name of the disease to compute impact for}

\item{outcome}{Name of the outcome (deaths, cases, dalys)}

\item{by_country}{Logical, indicating if impact should be computed by country}

\item{cohorts}{Vector of cohorts (default is all GAVI years)}

\item{present}{Year to interpret as present}
}
\description{
Create parameters for impact calculations
}
