% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_meta.R
\name{recipe_template}
\alias{recipe_template}
\title{Impact recipe template}
\usage{
recipe_template(template_version = "201710", method)
}
\arguments{
\item{template_version}{version can be any VIMC model run - e.g. 201710, 201910}

\item{method}{method can be any VIMC impact methods - method0, method1, method2a, method2b}
}
\description{
Generate impact recipe template
}
