% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpact.R
\name{split_scenario_vaccine_delivery}
\alias{split_scenario_vaccine_delivery}
\title{Take a string describing scenario & vaccine delivery and return in list form}
\usage{
split_scenario_vaccine_delivery(string)
}
\arguments{
\item{string}{String from a recipe describing a scenario type and vaccine
delivery method e.g. default:YF-routine;YF-campaign}
}
\value{
List representation of scenario type and vaccine delivery method
  as required by calculate_impact. e.g.
  list(
    scenario_type = "default",
    vaccine_delivery = list(
      list(
        vaccine = "YF",
        activity_type = "routine"
      ),
      list(
        vaccine = "YF",
        activity_type = "campaign"
      )
  )
}
\description{
Take a string of scenario and vaccine delivery from recipe and convert
into list representation for calling calculate_impact. If this
is a novac scenario will return
list(
  scenario_type = "novac",
  vaccine_delivery = NULL
)
}
\keyword{internal}
