% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonly_used_db_data.R
\name{extract_vaccination_history}
\alias{extract_vaccination_history}
\title{Generate target, fvps, coverage for a touchstone}
\usage{
extract_vaccination_history(
  con,
  touchstone_cov = "201710gavi",
  touchstone_pop = NULL,
  year_min = 2000,
  year_max = 2100,
  vaccine_to_ignore = c("DTP3", "HepB_BD_home"),
  gavi_support_levels = c("with", "bestminus")
)
}
\arguments{
\item{con}{Datebase connection}
}
\description{
Replace jenner:::fix_covreage_fvps() function
}
