% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commonly_used_db_data.R
\name{extract_vaccination_history}
\alias{extract_vaccination_history}
\title{Generate target, fvps, coverage for a touchstone}
\usage{
extract_vaccination_history(
  con,
  touchstone_cov = "201710gavi",
  touchstone_pop = NULL,
  year_min = 2000,
  year_max = 2100,
  vaccine_to_ignore = c("DTP3", "HepB_BD_home", "none"),
  disease_to_extract = NULL,
  countries_to_extract = NULL,
  gavi_support_levels = c("with", "bestminus"),
  scenario_type = "default",
  external_population_estimates = NULL,
  full_description = TRUE
)
}
\arguments{
\item{con}{Datebase connection}

\item{touchstone_cov}{Coverage touchstone}

\item{touchstone_pop}{Demography touchstone}

\item{year_min}{extract data from year_min}

\item{year_max}{extract data to year_max}

\item{vaccine_to_ignore}{Ignore defined vaccines}

\item{disease_to_extract}{extract data for specific diseases}

\item{countries_to_extract}{extract data for specific countries}

\item{gavi_support_levels}{gavi support levels}

\item{scenario_type}{scenario type}

\item{external_population_estimates}{The rationales are 1. we can use external population estimates if any and if necessary;
2. demographic uncertainty not only affects models, but also FVPs. If we are to conduct sensitivity analysis on impact_by_year_of_vaccination, we need to vary population input for adjusting FVPs.}

\item{full_description}{TRUE if including scenario_descriptions (coverage estiamtes will be duplicated for scenarios); and FALSE if only providing coverage estimates}
}
\description{
Replace jenner:::fix_covreage_fvps() function
}
