% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vimpact.R
\name{get_coverage_data}
\alias{get_coverage_data}
\title{Get coverage data for a particular touchstone and vaccine delivery method}
\usage{
get_coverage_data(
  con,
  touchstone,
  delivery_methods,
  countries = NULL,
  vaccination_years = NULL
)
}
\arguments{
\item{con}{DB connection}

\item{touchstone}{Touchstone to get data for}

\item{countries}{Optional vector of countries to filter data}

\item{vaccination_years}{Option vector of years to filter data}

\item{delivery}{List of delivery methods to get data for}
}
\value{
Tibble of coverage data
}
\description{
Get coverage data for a particular touchstone and vaccine delivery method
}
\keyword{internal}
