% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/impact_central.R
\name{impact_by_year_of_vaccination_activity_type}
\alias{impact_by_year_of_vaccination_activity_type}
\title{Calculate impact by year of vaccination: activity type}
\usage{
impact_by_year_of_vaccination_activity_type(
  baseline_impact,
  focal_impact,
  fvps,
  vaccination_years
)
}
\arguments{
\item{baseline_impact}{Data frame of baseline impact data this needs to have
columns country, burden_outcome, vaccine_delivery, year, age, value}

\item{focal_impact}{Data frame of focal impact data this needs to have
columns country, burden_outcome, vaccine_delivery, year, age, value}

\item{fvps}{Data frame of FVPs (fully vaccinated persons) needs to have
columns country, year, activity_type, fvps other columns can be included
and will be aggregated over.}

\item{vaccination_years}{Years of vaccination of interest.}
}
\value{
Vaccine impact by country, activity type (routine or campaign), year
and burden outcome
}
\description{
Impact by year of vaccination with impact ratio stratified by activity type.
Stratifying impact ratio by activity type captures the differing effects of
routine and campaign vaccination.
}
\details{
To calculate impact by year of vaccination using impact ratios stratified by
activity type, we assume that routine vaccination and campaign vaccination,
which target multiple age groups, have different effects; for example due
to dosage clustering. Hence, this method produces multiple,
activity-specific impact ratios which we then multiply by the number of
FVPs (fully vaccinated persons) to calculate impact.
}
